+++ packages/web-components/@sfx/autocomplete/src/events.ts
@@ -1,1 +1,1 @@
-export const autocompleteReceivedResults = 'sfx::autocomplete_received_results';
+export const AUTOCOMPLETE_RECEIVED_RESULTS_EVENT = 'sfx::autocomplete_received_results';